ISBN Check Exceptions

+++ src/controller.js
@@ -56,6 +56,11 @@ class LibrosController{
       if(!carga.id){
         throw new Error("ID Requerido");
       }
+
+      if(!carga.nombre || !carga.autor || !carga.categoria || !carga.año || !carga.ISBN){
+        throw new Error("Datos Requeridos");
+      }
+
       const [result] = await pool.query(`UPDATE libros SET nombre = (?), autor=(?), categoria=(?), \`año-publicacion\` = (?), ISBN= (?) WHERE id = (?)`, [carga.nombre,carga.autor,carga.categoria,carga.año,carga.ISBN,carga.id]);
       res.json({"Registros modificados":result.affectedRows});
     } catch (e) {
@@ -65,21 +70,49 @@ class LibrosController{
   }
 
   async getISBN(req,res){
-    const carga = req.body;
-    const [result] = await pool.query(`SELECT * FROM libros WHERE ISBN=(?)`,[carga.ISBN]);
-    res.json(result);
+    try {
+      const carga = req.body;
+      if(!carga.ISBN){
+        throw new Error("ISBN Requerido");
+      }
+      const [result] = await pool.query(`SELECT * FROM libros WHERE ISBN=(?)`,[carga.ISBN]);
+      res.json(result);
+    } catch (e) {
+      res.json({"Error":e.message});
+      console.log("Error:" + e.message);
+    }
   }
 
   async rmvISBN(req,res){
-    const carga = req.body;
-    const [result] = await pool.query(`DELETE FROM libros WHERE ISBN =(?)`,[carga.ISBN]);
-    res.json({"Registros eliminados":result.affectedRows});
+    try {
+      const carga = req.body;
+      if(!carga.ISBN){
+        throw new Error("ISBN Requerido");
+      }
+      const [result] = await pool.query(`DELETE FROM libros WHERE ISBN =(?)`,[carga.ISBN]);
+      res.json({"Registros eliminados":result.affectedRows});
+    } catch (e) {
+      res.json({"Error":e.message});
+      console.log("Error:" + e.message);
+    }
   }
 
   async mdfISBN(req, res){
-    const carga = req.body;
-    const [result] = await pool.query(`UPDATE libros SET nombre = (?), autor=(?), categoria=(?), \`año-publicacion\` = (?) WHERE ISBN = (?)`, [carga.nombre,carga.autor,carga.categoria,carga.año,carga.ISBN]);
-    res.json({"Registros modificados":result.affectedRows});
+    try {
+      const carga = req.body;
+      if(!carga.ISBN){
+        throw new Error("ISBN Requerido");
+      }
+      if(!carga.nombre || !carga.autor || !carga.categoria || !carga.año){
+        throw new Error("Datos Requeridos");
+      }
+
+      const [result] = await pool.query(`UPDATE libros SET nombre = (?), autor=(?), categoria=(?), \`año-publicacion\` = (?) WHERE ISBN = (?)`, [carga.nombre,carga.autor,carga.categoria,carga.año,carga.ISBN]);
+      res.json({"Registros modificados":result.affectedRows});
+    } catch (e){
+      res.json({"Error":e.message});
+      console.log("Error:" + e.message);
+    }
   }
